updated file structure | moved get_unverified_users() from Users model to VerifiedUsers model
